Introduce first-class `image` module in `iced_native`
