+++ glow/src/backend.rs
@@ -237,10 +237,7 @@ impl backend::Text for Backend {
 
 #[cfg(feature = "image")]
 impl backend::Image for Backend {
-    fn dimensions(
-        &self,
-        _handle: &iced_native::widget::image::Handle,
-    ) -> (u32, u32) {
+    fn dimensions(&self, _handle: &iced_native::image::Handle) -> (u32, u32) {
         (50, 50)
     }
 }
