+++ graphics/src/primitive.rs
@@ -1,4 +1,4 @@
-use iced_native::widget::image;
+use iced_native::image;
 use iced_native::widget::svg;
 use iced_native::{Background, Color, Font, Rectangle, Size, Vector};
