+++ native/src/lib.rs
@@ -36,6 +36,7 @@
 pub mod clipboard;
 pub mod command;
 pub mod event;
+pub mod image;
 pub mod keyboard;
 pub mod layout;
 pub mod mouse;
