+++ native/src/widget/image.rs
@@ -2,15 +2,12 @@
 pub mod viewer;
 pub use viewer::Viewer;
 
+use crate::image::{self, Handle};
 use crate::layout;
 use crate::renderer;
 use crate::{Element, Hasher, Layout, Length, Point, Rectangle, Size, Widget};
 
-use std::{
-    hash::{Hash, Hasher as _},
-    path::PathBuf,
-    sync::Arc,
-};
+use std::hash::Hash;
 
 /// A frame that displays an image while keeping aspect ratio.
 ///
@@ -55,7 +52,7 @@ impl Image {
 
 impl<Message, Renderer> Widget<Message, Renderer> for Image
 where
-    Renderer: self::Renderer,
+    Renderer: image::Renderer,
 {
     fn width(&self) -> Length {
         self.width
@@ -111,129 +108,9 @@ where
     }
 }
 
-/// An [`Image`] handle.
-#[derive(Debug, Clone)]
-pub struct Handle {
-    id: u64,
-    data: Arc<Data>,
-}
-
-impl Handle {
-    /// Creates an image [`Handle`] pointing to the image of the given path.
-    ///
-    /// Makes an educated guess about the image format by examining the data in the file.
-    pub fn from_path<T: Into<PathBuf>>(path: T) -> Handle {
-        Self::from_data(Data::Path(path.into()))
-    }
-
-    /// Creates an image [`Handle`] containing the image pixels directly. This
-    /// function expects the input data to be provided as a `Vec<u8>` of BGRA
-    /// pixels.
-    ///
-    /// This is useful if you have already decoded your image.
-    pub fn from_pixels(width: u32, height: u32, pixels: Vec<u8>) -> Handle {
-        Self::from_data(Data::Pixels {
-            width,
-            height,
-            pixels,
-        })
-    }
-
-    /// Creates an image [`Handle`] containing the image data directly.
-    ///
-    /// Makes an educated guess about the image format by examining the given data.
-    ///
-    /// This is useful if you already have your image loaded in-memory, maybe
-    /// because you downloaded or generated it procedurally.
-    pub fn from_memory(bytes: Vec<u8>) -> Handle {
-        Self::from_data(Data::Bytes(bytes))
-    }
-
-    fn from_data(data: Data) -> Handle {
-        let mut hasher = Hasher::default();
-        data.hash(&mut hasher);
-
-        Handle {
-            id: hasher.finish(),
-            data: Arc::new(data),
-        }
-    }
-
-    /// Returns the unique identifier of the [`Handle`].
-    pub fn id(&self) -> u64 {
-        self.id
-    }
-
-    /// Returns a reference to the image [`Data`].
-    pub fn data(&self) -> &Data {
-        &self.data
-    }
-}
-
-impl<T> From<T> for Handle
-where
-    T: Into<PathBuf>,
-{
-    fn from(path: T) -> Handle {
-        Handle::from_path(path.into())
-    }
-}
-
-impl Hash for Handle {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        self.id.hash(state);
-    }
-}
-
-/// The data of an [`Image`].
-#[derive(Clone, Hash)]
-pub enum Data {
-    /// File data
-    Path(PathBuf),
-
-    /// In-memory data
-    Bytes(Vec<u8>),
-
-    /// Decoded image pixels in BGRA format.
-    Pixels {
-        /// The width of the image.
-        width: u32,
-        /// The height of the image.
-        height: u32,
-        /// The pixels.
-        pixels: Vec<u8>,
-    },
-}
-
-impl std::fmt::Debug for Data {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Data::Path(path) => write!(f, "Path({:?})", path),
-            Data::Bytes(_) => write!(f, "Bytes(...)"),
-            Data::Pixels { width, height, .. } => {
-                write!(f, "Pixels({} * {})", width, height)
-            }
-        }
-    }
-}
-
-/// The renderer of an [`Image`].
-///
-/// Your [renderer] will need to implement this trait before being able to use
-/// an [`Image`] in your user interface.
-///
-/// [renderer]: crate::renderer
-pub trait Renderer: crate::Renderer {
-    /// Returns the dimensions of an [`Image`] located on the given path.
-    fn dimensions(&self, handle: &Handle) -> (u32, u32);
-
-    // Draws an [`Image`].
-    fn draw(&mut self, handle: Handle, bounds: Rectangle);
-}
-
 impl<'a, Message, Renderer> From<Image> for Element<'a, Message, Renderer>
 where
-    Renderer: self::Renderer,
+    Renderer: image::Renderer,
 {
     fn from(image: Image) -> Element<'a, Message, Renderer> {
         Element::new(image)
