+++ src/widget.rs
@@ -38,8 +38,9 @@ mod platform {
     #[cfg_attr(docsrs, doc(cfg(feature = "image")))]
     pub mod image {
         //! Display images in your user interface.
+        pub use crate::runtime::image::Handle;
         pub use crate::runtime::widget::image::viewer;
-        pub use crate::runtime::widget::image::{Handle, Image, Viewer};
+        pub use crate::runtime::widget::image::{Image, Viewer};
     }
 
     #[cfg_attr(docsrs, doc(cfg(feature = "svg")))]
